Move /auth/export code to own file
